pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-sql"
version = "2.0.0"
description = "Best-in-class agentic text-to-SQL framework with ReAct, Reflection, and Multi-Agent patterns"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Agentic SQL Team" },
]
keywords = [
    "text-to-sql",
    "llm",
    "ai-agents",
    "natural-language",
    "database",
    "rag",
    "memory",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=2.16.0",
    "pydantic>=2.0.0",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
all-llms = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "sse-starlette>=1.6.0",
]
auth = [
    "PyJWT>=2.8.0",
]
postgres = [
    "asyncpg>=0.28.0",
    "psycopg2-binary>=2.9.0",
]
mysql = [
    "aiomysql>=0.2.0",
]
sqlite = [
    "aiosqlite>=0.19.0",
]
snowflake = [
    "snowflake-connector-python>=3.0.0",
]
bigquery = [
    "google-cloud-bigquery>=3.0.0",
]
vector = [
    "chromadb>=0.4.0",
    "pgvector>=0.2.0",
]
graph = [
    "neo4j>=5.0.0",
]
memory = [
    "chromadb>=0.4.0",
    "neo4j>=5.0.0",
]
viz = [
    "plotly>=5.0.0",
    "pandas>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "agentic-sql[all-llms,api,auth,postgres,vector,graph,viz,dev]",
]

[project.urls]
Homepage = "https://github.com/agentic-sql/agentic-sql"
Documentation = "https://agentic-sql.readthedocs.io"
Repository = "https://github.com/agentic-sql/agentic-sql"

[project.scripts]
agentic-sql = "agentic_sql.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
